modified signal handler.  it aborts when it is attempted to be killed twice.  not initialize system before starting interactive session on UNIX.
git-svn-id: svn+ssh://savannah.psi.ch/afs/psi.ch/project/meg/svn/rome/trunk/rome@1298 306b0a18-5003-0410-92f3-cc1c673a7936

+++ src/ROMERint.cpp
@@ -32,6 +32,8 @@ public:
 
 Bool_t ROMEInterruptHandler::Notify()
 {
+   static Int_t nNotified = 0;
+   nNotified++;
    gROME->ss_getchar(1);
    gROME->restoreOutput();
    if (strcmp(gROME->GetActiveDAQ()->GetName(), "midas") == 0
@@ -42,7 +44,11 @@ Bool_t ROMEInterruptHandler::Notify()
    }
    if (gROME->GetApplication()->isUseRintInterruptHandler())
       return gROME->GetApplication()->GetRintInterruptHandler()->Notify();
-   gROME->GetApplication()->Terminate(0);
+
+   if (nNotified >= 2)
+      gSystem->Abort();
+   else
+      gROME->GetApplication()->Terminate(0);
    return kTRUE;
 }
